Add files via upload
This is a Tic Tac Toe game separated into several functions. and files. I used a 2d array as a map and several loops to check for valid user input. the game has a single player and a multi player mode. in the first case the opponent is a very rudimentary random number generator.